Split existing tests between files
