+++ cypress/integration/pages/articles/testsForCanonicalOnly.js
@@ -1,0 +1,72 @@
+import envConfig from '../../../support/config/envs';
+import config from '../../../support/config/services';
+
+// TODO: Remove after https://github.com/bbc/simorgh/issues/2959
+const serviceHasCaption = service => service === 'news';
+
+const tests = ({ service }) =>
+  describe(`Canonical Tests`, () => {
+    it('should not have an AMP attribute on the main article', () => {
+      cy.get('html').should('not.have.attr', 'amp');
+    });
+
+    describe('ATI', () => {
+      it('should have a noscript tag with an 1px image with the ati url', () => {
+        cy.hasNoscriptImgAtiUrl(
+          envConfig.atiUrl,
+          config[service].isWorldService ? envConfig.atiAnalyticsWSBucket : '',
+        );
+      });
+    });
+
+    describe('Chartbeat', () => {
+      if (envConfig.chartbeatEnabled) {
+        it('should have a script with src value set to chartbeat source', () => {
+          cy.hasScriptWithChartbeatSrc();
+        });
+        it('should have chartbeat config set to window object', () => {
+          cy.hasGlobalChartbeatConfig();
+        });
+      }
+    });
+
+    it('should include ampHTML tag', () => {
+      cy.get('head link[rel="amphtml"]').should(
+        'have.attr',
+        'href',
+        `${window.location.origin}${config[service].pageTypes.articles.path}.amp`,
+      );
+    });
+
+    if (serviceHasCaption(service)) {
+      it('should have a visible image with a caption that is lazyloaded and has a noscript fallback image', () => {
+        cy.get('figure')
+          .eq(2)
+          .within(() => {
+            cy.get('div div div div').should(
+              'have.class',
+              'lazyload-placeholder',
+            );
+          })
+          .scrollIntoView();
+
+        cy.get('figure')
+          .eq(2)
+          .should('be.visible')
+          .should('to.have.descendants', 'img')
+          .should('to.have.descendants', 'figcaption')
+
+          // NB: If this test starts failing unexpectedly it's a good sign that the dom is being
+          // cleared during hydration. React won't render noscript tags on the client so if they
+          // get cleared during hydration, the following render wont re-add them.
+          // See https://github.com/facebook/react/issues/11423#issuecomment-341751071 or
+          // https://github.com/bbc/simorgh/pull/1872 for more infomation.
+          .within(() => {
+            cy.get('noscript').contains('<img ');
+            cy.get('div div').should('not.have.class', 'lazyload-placeholder');
+          });
+      });
+    }
+  });
+
+export default tests;
