+++ cypress/integration/pages/frontPage/tests.js
@@ -1,0 +1,48 @@
+const tests = () =>
+  describe(`Tests`, () => {
+    describe('Frontpage body', () => {
+      before(() => {
+        cy.viewport(1008, 768);
+      });
+
+      describe('Header', () => {
+        it('should have a visually hidden top-level header', () => {
+          cy.get('h1').should('have.length', 1);
+        });
+      });
+
+      describe('Section', () => {
+        it('should be labelled by a visible section label', () => {
+          cy.get('section')
+            .should('have.length.of.at.least', 1)
+            .should('be.visible')
+            .each($section => {
+              cy.wrap($section).within(() => {
+                cy.get('h2').should('have.lengthOf', 1);
+              });
+            });
+        });
+
+        it('should contain at least one story promo', () => {
+          cy.get('section').within(() => {
+            cy.get('img')
+              .should('have.length.of.at.least', 1)
+              .should('be.visible');
+            cy.get('h3')
+              .should('have.length.of.at.least', 1)
+              .should('be.visible')
+              .find('a')
+              .should('have.attr', 'href');
+            cy.get('p')
+              .should('have.length.of.at.least', 1)
+              .should('be.visible');
+            cy.get('time')
+              .should('have.length.of.at.least', 1)
+              .should('be.visible');
+          });
+        });
+      });
+    });
+  });
+
+export default tests;
