+++ cypress/support/helpers/runTestsForPage.js
@@ -15,7 +15,7 @@ const runTestsForPage = (
 ) => {
   Object.keys(config)
     .filter(service => serviceHasPageType(service, pageType))
-    .filter(service => shouldSmokeTest(pageType, service))
+    .filter(service => shouldSmokeTest(pageType, service)) // I cannot do this and override
     .forEach(service => {
       describe(`${pageType} - ${service} - Canonical`, () => {
         before(() => {
